Split Bitfinex streaming services
